# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WhatsApp Clone API",
        "description": "API documentation for WhatsApp Clone - Real-time messaging application",
        "contact": {
            "email": "support@whatsappclone.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/files/upload": {
            "post": {
                "tags": [
                    "Files"
                ],
                "summary": "Upload a file to a chat",
                "description": "Upload a file to a chat.",
                "operationId": "df81c005643f626c399f24bf22fc4976",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/FileUploadRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "File uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/FileUploadResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - User is not a participant of the chat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error - Invalid file type, size, or format",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/files/download/{message}": {
            "get": {
                "tags": [
                    "Files"
                ],
                "summary": "Download a file from a message",
                "description": "Download a file from a message.",
                "operationId": "691d44cdb00139e732083de185594e5e",
                "parameters": [
                    {
                        "name": "message",
                        "in": "path",
                        "description": "Message ID containing the file",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File downloaded successfully",
                        "headers": {
                            "Content-Disposition": {
                                "description": "Attachment with filename",
                                "schema": {
                                    "type": "string",
                                    "example": "attachment; filename='document.pdf'"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - User is not a participant of the chat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Validation error"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": [
                            {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "property": null
                                },
                                "property": null
                            }
                        ]
                    }
                },
                "type": "object"
            },
...
            "FileUploadRequest": {
                "required": [
                    "file",
                    "type",
                    "chat_id"
                ],
                "properties": {
                    "file": {
                        "description": "The file to upload (max 5MB)",
                        "type": "string",
                        "format": "binary"
                    },
                    "type": {
                        "description": "Type of file being uploaded",
                        "type": "string",
                        "enum": [
                            "image",
                            "document",
                            "audio"
                        ]
                    },
                    "chat_id": {
                        "description": "ID of the chat to upload the file to",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "FileUploadResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "content": {
                        "type": "string",
                        "example": "document.pdf"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 123
                    },
                    "chat_id": {
                        "type": "integer",
                        "example": 456
                    },
                    "message_type": {
                        "type": "string",
                        "enum": [
                            "image",
                            "document",
                            "audio"
                        ],
                        "example": "document"
                    },
                    "file_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/storage/uploads/documents/document.pdf"
                    },
                    "file_name": {
                        "type": "string",
                        "example": "document.pdf"
                    },
                    "file_size": {
                        "type": "integer",
                        "example": 1024000
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 123
                            },
                            "name": {
                                "type": "string",
                                "example": "John Doe"
                            },
                            "avatar": {
                                "type": "string",
                                "example": "https://example.com/avatar.jpg"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}